% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate_SEEE_indicator.R
\name{initiate_SEEE_indicator}
\alias{initiate_SEEE_indicator}
\title{Initiate a SEEE indicator}
\usage{
initiate_SEEE_indicator(
  indic,
  type = "Outil d'évaluation",
  author = "",
  version = "1.0.0"
)
}
\arguments{
\item{indic}{string. The name of the indicator}

\item{type}{string. The type of indicator (e.g. Outil d'évaluation, Outil de
diagnostic, beta-test)}

\item{author}{string. The names of the indicator authors}

\item{set_renv}{boolean. If TRUE, will set renv up installing the packages
available on the SEEE server with the same version.}
}
\value{
nothing. Writes in the current directory the template files necessary
  to develop a SEEE indicator
}
\description{
This function allows to generate the template files necessary to develop a
SEEE indicator. Basic information about indicator name, type and authors are
added to the files.
}
