% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_renv.R
\name{set_renv}
\alias{set_renv}
\title{Set up renv packages}
\usage{
set_renv(package_list = NULL)
}
\arguments{
\item{package_list}{a data frame (or a named list) with two columns
(elements): Package containing the name of the packages to install and
Version containing the specific version to install (if no specific version
required use "")}
}
\value{

}
\description{
Set up a project library and install the specified (by default SEEE server)
package list
}
