% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_SEEE_export.R
\name{create_SEEE_export}
\alias{create_SEEE_export}
\title{Create a SEEE export}
\usage{
create_SEEE_export(indic, additionalInput = NULL)
}
\arguments{
\item{indic}{the name of the indicator as it is in the names of the
development files (ending with _valid.r and _calc.r)}
}
\description{
Create the folder and files necessary to upload to the SEEE server in order
to implement a new indicator
}
